+++ server/src/main/java/seb15/roobits/member/entity/Member.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import seb15.roobits.member.Auditable.Auditable;
+import seb15.roobits.room.entity.Room;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
@@ -53,15 +54,15 @@ public class Member extends Auditable {
 
 
 //    룸부분 결합후 작업
-//    @OneToMany(mappedBy = "Room")
-//    private List<Room> rooms = new ArrayList<>();
-//
-//    public void setRoom(Room room){
-//        rooms.add(room);
+    @OneToMany(mappedBy = "Room")
+    private List<Room> rooms = new ArrayList<>();
+
+    public void setRoom(Room room){
+        rooms.add(room);
 //        if(room.getRoom() != this){
 //            room.setRoom(this);
 //        }
-//    }
+    }
 
     public List<String> getRoleList(){
         if(this.roles != null){
