+++ server/src/main/java/seb15/roobits/room/repository/RoomRepository.java
@@ -10,6 +10,6 @@ public interface RoomRepository extends JpaRepository<Room, Long> {
 
     Optional<Room> findById(long boardId);
 
-//    @Query("SELECT * FROM ROOM WHERE ROOM_NAME = :roomName")
+    @Query("SELECT * FROM ROOM WHERE ROOM_NAME = :roomName")
     Optional<Room> findByRoomName(String roomName);
 }
